requirements.txt
django===2.1.1
django-cors-headers===2.4.0
django-environ==0.4.5
djangorestframework==3.7.7
Pillow==5.2.0
psycopg2==2.7.4
python-decouple
python-magic==0.4.15
requests==2.18.4
